backend/requirements.txt
Django>=6.0
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg2-binary>=2.9
djangorestframework-simplejwt>=5.3
python-dotenv>=1.0
